+++ migrations/mysql/2026-07-26-000001_add_users_organizations_external_id_index/down.sql
@@ -1,1 +1,0 @@
-DROP INDEX idx_users_organizations_org_external_id ON users_organizations;