+++ migrations/mysql/2026-07-26-000002_add_groups_external_id_index/down.sql
@@ -1,1 +1,0 @@
-DROP INDEX idx_groups_org_external_id ON `groups`;