+++ migrations/mysql/2026-08-09-000001_drop_redundant_users_organizations_external_id_index/down.sql
@@ -1,2 +1,0 @@
-CREATE INDEX idx_users_organizations_org_external_id
-	ON users_organizations (org_uuid, external_id(150));