+++ migrations/mysql/2026-08-09-000002_drop_redundant_groups_external_id_index/down.sql
@@ -1,2 +1,0 @@
-CREATE INDEX idx_groups_org_external_id
-	ON `groups` (organizations_uuid, external_id(150));