+++ migrations/sqlite/2026-08-08-000002_unique_groups_external_id/down.sql
@@ -1,1 +1,0 @@
-DROP INDEX IF EXISTS idx_groups_org_external_id_unique;